表结构大变动，主键位数升级，对应实体类由Integer升级成Long 顺便梳理实体类

+++ cdp-cloudadviced/cdp-platform/cdp-platformapi/src/main/java/com/github/fosin/cdp/platformapi/constant/SystemConstant.java
@@ -11,8 +11,8 @@ public class SystemConstant {
     public static final String ADMIN_USER_CODE = "admin";
     public static final String SUPER_ROLE_NAME = "SUPER";
     public static final String ADMIN_ROLE_NAME = "ADMIN";
-    public static final Integer ROOT_PERMISSION_PID = 0;
-    public static final Integer ROOT_PERMISSION_ID = 1;
+    public static final Long ROOT_PERMISSION_PID = 0L;
+    public static final Long ROOT_PERMISSION_ID = 1L;
     public static final Integer SYSTEM_DICTIONARY_TYPE = 1;
 
 }
