+++ cdp-cloudadviced/cdp-platform/cdp-platformapi/src/main/java/com/github/fosin/cdp/platformapi/repository/DictionaryRepository.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
  * @author fosin
  */
 @Repository
-public interface DictionaryRepository extends JpaRepository<CdpSysDictionaryEntity,Integer>,
+public interface DictionaryRepository extends JpaRepository<CdpSysDictionaryEntity,Long>,
                                     JpaSpecificationExecutor<CdpSysDictionaryEntity> {
 }
