+++ cdp-cloudadviced/cdp-platform/cdp-platformapi/src/main/java/com/github/fosin/cdp/platformapi/repository/PermissionRepository.java
@@ -15,7 +15,7 @@ import java.util.List;
  * @author fosin
  */
 @Repository
-public interface PermissionRepository extends JpaRepository<CdpSysPermissionEntity, Integer>, JpaSpecificationExecutor<CdpSysPermissionEntity> {
+public interface PermissionRepository extends JpaRepository<CdpSysPermissionEntity, Long>, JpaSpecificationExecutor<CdpSysPermissionEntity> {
     //    @Query(value = "select menu.* from cdp_sys_menu menu,cdp_sys_role_privilege p where menu.id=p.menu_id and p.role_id=?1",nativeQuery = true)
 //    List<CdpSysPermissionEntity> findRoleMenuByRoleId(Integer roleId);
 //
@@ -24,7 +24,7 @@ public interface PermissionRepository extends JpaRepository<CdpSysPermissionEnti
 //
 //    @Query(value = "select m.* from cdp_sys_menu m,(select p.role_id,p.menu_id,r.add_mode from cdp_sys_role_privilege p left join cdp_sys_user_privilege r on p.role_id =r.role_id and p.menu_id=r.menu_id  and r.user_id=?1  where p.role_id=?2 and (r.add_mode = 0 or r.add_mode is null)) z where m.id=z.menu_id union select m.* from cdp_sys_menu m ,cdp_sys_user_privilege r where m.id=r.menu_id and r.role_id=?2 and r.user_id=?1 and r.add_mode=0",nativeQuery = true)
 //    List<CdpSysPermissionEntity> getAllMenuByUserIdAndRoleId(Integer userId, Integer roleId);
-    List<CdpSysPermissionEntity> findByPId(Integer pId, Sort sort);
+    List<CdpSysPermissionEntity> findByPId(Long pId, Sort sort);
 
     List<CdpSysPermissionEntity> findByType(Integer type);
