+++ cdp-cloudadviced/cdp-platform/cdp-platformapi/src/main/java/com/github/fosin/cdp/platformapi/repository/UserRepository.java
@@ -14,13 +14,13 @@ import java.util.List;
  * @author fosin
  */
 @Repository
-public interface UserRepository extends JpaRepository<CdpSysUserEntity,Integer>,JpaSpecificationExecutor<CdpSysUserEntity> {
+public interface UserRepository extends JpaRepository<CdpSysUserEntity,Long>,JpaSpecificationExecutor<CdpSysUserEntity> {
     CdpSysUserEntity findByUsercode(String usercode);
 
     @Query(value = "select * from cdp_sys_user where id not in (select user_id from cdp_sys_user_role where role_id =?1)",nativeQuery = true)
-    List<CdpSysUserEntity> findOtherUsersByRoleId(Integer roleId);
+    List<CdpSysUserEntity> findOtherUsersByRoleId(Long roleId);
 
     @Query(value = "select * from cdp_sys_user where id in (select user_id from cdp_sys_user_role where role_id =?1)",nativeQuery = true)
-    List<CdpSysUserEntity> findRoleUsersByRoleId(Integer roleId);
+    List<CdpSysUserEntity> findRoleUsersByRoleId(Long roleId);
 }
