+++ cdp-cloudadviced/cdp-platform/cdp-platformapi/src/main/java/com/github/fosin/cdp/platformapi/service/UserPermissionServiceImpl.java
@@ -1,12 +1,12 @@
 package com.github.fosin.cdp.platformapi.service;
 
-import com.github.fosin.cdp.platformapi.repository.UserPermissionRepository;
+import com.github.fosin.cdp.cache.util.CacheUtil;
 import com.github.fosin.cdp.core.exception.CdpServiceException;
 import com.github.fosin.cdp.platformapi.constant.TableNameConstant;
 import com.github.fosin.cdp.platformapi.entity.CdpSysUserEntity;
 import com.github.fosin.cdp.platformapi.entity.CdpSysUserPermissionEntity;
+import com.github.fosin.cdp.platformapi.repository.UserPermissionRepository;
 import com.github.fosin.cdp.platformapi.service.inter.IUserPermissionService;
-import com.github.fosin.cdp.cache.util.CacheUtil;
 import com.github.fosin.cdp.platformapi.util.LoginUserUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
@@ -31,18 +31,18 @@ public class UserPermissionServiceImpl implements IUserPermissionService {
 
     @Override
     @Cacheable(value = TableNameConstant.CDP_SYS_USER_PERMISSION, key = "T(String).valueOf(#userId).concat('-').concat(T(String).valueOf(#organizId))")
-    public List<CdpSysUserPermissionEntity> findByUserIdAndOrganizId(Integer userId, Integer organizId) {
+    public List<CdpSysUserPermissionEntity> findByUserIdAndOrganizId(Long userId, Long organizId) {
         return userPermissionRepository.findByUserIdAndOrganizId(userId, organizId);
     }
 
     @Override
     @Cacheable(value = TableNameConstant.CDP_SYS_USER_PERMISSION, key = "#userId")
-    public List<CdpSysUserPermissionEntity> findByUserId(Integer userId) {
+    public List<CdpSysUserPermissionEntity> findByUserId(Long userId) {
         return userPermissionRepository.findByUserId(userId);
     }
 
     @Override
-    public long countByPermissionId(Integer permissionId) {
+    public long countByPermissionId(Long permissionId) {
         return userPermissionRepository.countByPermissionId(permissionId);
     }
 
@@ -59,12 +59,12 @@ public class UserPermissionServiceImpl implements IUserPermissionService {
 
     @Override
     public void deleteInBatch(Collection<CdpSysUserPermissionEntity> entities) throws CdpServiceException {
-        Set<Integer> needDelUsers = new HashSet<>();
+        Set<Long> needDelUsers = new HashSet<>();
         for (CdpSysUserPermissionEntity entity : entities) {
             needDelUsers.add(entity.getUserId());
         }
         Assert.isTrue(needDelUsers.size() != 0, "没有找到需要删除数据!");
-        for (Integer userId : needDelUsers) {
+        for (Long userId : needDelUsers) {
             CacheUtil.evict(TableNameConstant.CDP_SYS_USER_PERMISSION, userId + "");
         }
         userPermissionRepository.deleteInBatch(entities);
@@ -73,7 +73,7 @@ public class UserPermissionServiceImpl implements IUserPermissionService {
         } catch (InterruptedException e) {
             throw new CdpServiceException(e);
         }
-        for (Integer userId : needDelUsers) {
+        for (Long userId : needDelUsers) {
             CacheUtil.evict(TableNameConstant.CDP_SYS_USER_PERMISSION, userId + "");
         }
     }
@@ -90,7 +90,7 @@ public class UserPermissionServiceImpl implements IUserPermissionService {
             }
     )
     @Transactional
-    public List<CdpSysUserPermissionEntity> updateInBatch(Integer userId, Iterable<CdpSysUserPermissionEntity> entities) {
+    public List<CdpSysUserPermissionEntity> updateInBatch(Long userId, Iterable<CdpSysUserPermissionEntity> entities) {
         Assert.notNull(userId, "传入的用户ID不能为空!");
         for (CdpSysUserPermissionEntity entity : entities) {
             Assert.isTrue(entity.getUserId().equals(userId),"需要更新的数据集中有与用户ID不匹配的数据!");
